Move DocumentNode
